requirements.txt
Flask
mysql-connector-python
firebase-admin
python-dotenv
gunicorn
